Refactor license formatting into a reusable utility and add license status to the support bundle. Closes #18207
